@@ -1,56 +1,43 @@
 {
     "name": "dragon-code/support",
     "description": "Support package is a collection of helpers and tools for any project.",
-    "license": "MIT",
     "type": "library",
-    "keywords": [
-        "support",
-        "laravel",
-        "symfony",
-        "yii",
-        "yii2",
-        "framework",
-        "helper",
-        "helpers",
-        "dragon-code",
-        "dragon"
-    ],
+    "license": "MIT",
+    "keywords": ["support", "laravel", "symfony", "yii", "yii2", "framework", "helper", "helpers", "dragon-code", "dragon"],
+    "support": {
+        "issues": "https://github.com/TheDragonCode/support/issues",
+        "source": "https://github.com/TheDragonCode/support"
+    },
     "authors": [
         {
             "name": "Andrey Helldar",
             "email": "helldar@ai-rus.com"
         }
     ],
-    "support": {
-        "issues": "https://github.com/TheDragonCode/support/issues",
-        "source": "https://github.com/TheDragonCode/support"
-    },
     "require": {
-        "php": "^7.2.5 || ^8.0",
+        "php": "^7.4|^8.0",
         "ext-ctype": "*",
         "ext-dom": "*",
         "ext-json": "*",
         "ext-mbstring": "*",
-        "dragon-code/contracts": "^2.15",
+        "dragon-code/contracts": "^2.14",
         "psr/http-message": "^1.0.1",
-        "symfony/deprecation-contracts": "^2.5 || ^3.0",
-        "voku/portable-ascii": "^1.4.8 || ^2.0"
+        "symfony/deprecation-contracts": "^2.5",
+        "voku/portable-ascii": "^1.4.8"
     },
     "require-dev": {
         "ext-bcmath": "*",
-        "phpunit/phpunit": "^8.0 || ^9.0",
-        "symfony/var-dumper": "^5.2 || ^6.0"
+        "symfony/var-dumper": "^5.2",
+        "phpunit/phpunit": "^8.0|^9.0"
     },
     "conflict": {
         "andrey-helldar/support": "*"
     },
     "suggest": {
-        "ext-bcmath": "Require the extension if you will be using DragonCode\\Support\\Facades\\Helpers\\Digit.",
         "dragon-code/laravel-support": "Various helper files for the Laravel and Lumen frameworks",
-        "symfony/thanks": "Give thanks (in the form of a GitHub) to your fellow PHP package maintainers"
+        "symfony/thanks": "Give thanks (in the form of a GitHub) to your fellow PHP package maintainers",
+        "ext-bcmath": "Require the extension if you will be using DragonCode\\Support\\Facades\\Helpers\\Digit."
     },
-    "minimum-stability": "stable",
-    "prefer-stable": true,
     "autoload": {
         "psr-4": {
             "DragonCode\\Support\\": "src"
@@ -64,5 +51,7 @@
     "config": {
         "preferred-install": "dist",
         "sort-packages": true
-    }
+    },
+    "minimum-stability": "stable",
+    "prefer-stable": true
 }
