@@ -79,9 +79,7 @@ class Arr
      */
     public function renameKeysMap(array $array, array $map): array
     {
-        return $this->renameKeys($array, static function ($key) use ($map) {
-            return $map[$key] ?? $key;
-        });
+        return $this->renameKeys($array, static fn ($key) => $map[$key] ?? $key);
     }
 
     /**
@@ -438,9 +436,7 @@ class Arr
     {
         $callback = is_callable($keys)
             ? $keys
-            : static function ($key) use ($keys) {
-                return empty($keys) || ! in_array($key, (array) $keys);
-            };
+            : static fn ($key) => empty($keys) || ! in_array($key, (array) $keys);
 
         return $this->filter((array) $array, $callback, ARRAY_FILTER_USE_KEY);
     }
@@ -788,9 +784,7 @@ class Arr
      */
     public function storeAsJson(string $path, $array, bool $sort_keys = false, int $flags = 0): void
     {
-        $this->prepareToStore($path, StubTool::JSON, $array, static function (array $array) use ($flags) {
-            return json_encode($array, $flags);
-        }, $sort_keys);
+        $this->prepareToStore($path, StubTool::JSON, $array, static fn (array $array) => json_encode($array, $flags), $sort_keys);
     }
 
     /**
@@ -802,9 +796,7 @@ class Arr
      */
     public function storeAsArray(string $path, $array, bool $sort_keys = false): void
     {
-        $this->prepareToStore($path, StubTool::PHP_ARRAY, $array, static function (array $array) {
-            return var_export($array, true);
-        }, $sort_keys);
+        $this->prepareToStore($path, StubTool::PHP_ARRAY, $array, static fn (array $array) => var_export($array, true), $sort_keys);
     }
 
     /**
