@@ -77,17 +77,11 @@ class ArrTest extends TestCase
             400   => 'Num 400',
         ];
 
-        $this->assertSame(['baz' => 'Baz', 200 => 'Num 200', 400 => 'Num 400'], $this->arr()->except($arr, static function ($key) {
-            return ! Str::startsWith($key, ['foo', 'bar']);
-        }));
+        $this->assertSame(['baz' => 'Baz', 200 => 'Num 200', 400 => 'Num 400'], $this->arr()->except($arr, static fn ($key): bool => ! Str::startsWith($key, ['foo', 'bar'])));
 
-        $this->assertSame(['foo' => 'Foo', 200 => 'Num 200', 400 => 'Num 400'], $this->arr()->except($arr, static function ($key) {
-            return ! Str::startsWith($key, 'ba');
-        }));
+        $this->assertSame(['foo' => 'Foo', 200 => 'Num 200', 400 => 'Num 400'], $this->arr()->except($arr, static fn ($key): bool => ! Str::startsWith($key, 'ba')));
 
-        $this->assertSame(['foo' => 'Foo', 'bar' => 'Bar', 'baz' => 'Baz'], $this->arr()->except($arr, static function ($key) {
-            return ! is_numeric($key);
-        }));
+        $this->assertSame(['foo' => 'Foo', 'bar' => 'Bar', 'baz' => 'Baz'], $this->arr()->except($arr, static fn ($key): bool => ! is_numeric($key)));
     }
 
     public function testRenameKeys()
@@ -110,13 +104,9 @@ class ArrTest extends TestCase
             'baz_789' => 789,
         ];
 
-        $renamed = $this->arr()->renameKeys($source, static function ($key) {
-            return mb_strtoupper($key);
-        });
+        $renamed = $this->arr()->renameKeys($source, static fn ($key): string => mb_strtoupper($key));
 
-        $modified = $this->arr()->renameKeys($source, static function ($key, $value) {
-            return mb_strtolower($key) . '_' . $value;
-        });
+        $modified = $this->arr()->renameKeys($source, static fn ($key, $value): string => mb_strtolower($key) . '_' . $value);
 
         $this->assertSame($expected_renamed, $renamed);
         $this->assertSame($expected_modified, $modified);
@@ -450,17 +440,11 @@ class ArrTest extends TestCase
             400   => 'Num 400',
         ];
 
-        $this->assertSame(['foo' => 'Foo', 'bar' => 'Bar'], $this->arr()->only($arr, static function ($key) {
-            return Str::startsWith($key, ['foo', 'bar']);
-        }));
+        $this->assertSame(['foo' => 'Foo', 'bar' => 'Bar'], $this->arr()->only($arr, static fn ($key): bool => Str::startsWith($key, ['foo', 'bar'])));
 
-        $this->assertSame(['bar' => 'Bar', 'baz' => 'Baz'], $this->arr()->only($arr, static function ($key) {
-            return Str::startsWith($key, 'ba');
-        }));
+        $this->assertSame(['bar' => 'Bar', 'baz' => 'Baz'], $this->arr()->only($arr, static fn ($key): bool => Str::startsWith($key, 'ba')));
 
-        $this->assertSame([200 => 'Num 200', 400 => 'Num 400'], $this->arr()->only($arr, static function ($key) {
-            return is_numeric($key);
-        }));
+        $this->assertSame([200 => 'Num 200', 400 => 'Num 400'], $this->arr()->only($arr, static fn ($key): bool => is_numeric($key)));
     }
 
     public function testFilter()
@@ -479,9 +463,10 @@ class ArrTest extends TestCase
             200   => 'Num 200',
         ];
 
-        $result = $this->arr()->filter($source, static function ($value, $key) {
-            return Str::contains($value, 200) || Str::startsWith($key, 'b');
-        }, ARRAY_FILTER_USE_BOTH);
+        $result = $this->arr()->filter($source,
+            static fn ($value, $key): bool => Str::contains($value, 200) || Str::startsWith($key, 'b'),
+            ARRAY_FILTER_USE_BOTH
+        );
 
         $this->assertSame($target, $result);
     }
@@ -1574,9 +1559,7 @@ class ArrTest extends TestCase
             ],
         ];
 
-        $this->assertSame($expected, $this->arr()->map($source, static function ($value, $key) {
-            return Str::studly($key) . '_' . ($value * 2);
-        }));
+        $this->assertSame($expected, $this->arr()->map($source, static fn ($value, $key): string => Str::studly($key) . '_' . ($value * 2)));
     }
 
     public function testMapRecursive()
@@ -1605,9 +1588,7 @@ class ArrTest extends TestCase
             ],
         ];
 
-        $this->assertSame($expected, $this->arr()->map($source, static function ($value, $key) {
-            return Str::studly($key) . '_' . ($value * 2);
-        }, true));
+        $this->assertSame($expected, $this->arr()->map($source, static fn ($value, $key): string => Str::studly($key) . '_' . ($value * 2), true));
     }
 
     public function testPush()
