@@ -29,7 +29,7 @@ use DragonCode\Support\Helpers\Call as Helper;
  */
 class Call extends Facade
 {
-    protected static function getFacadeAccessor()
+    protected static function getFacadeAccessor(): string
     {
         return Helper::class;
     }
