@@ -28,7 +28,7 @@ use DragonCode\Support\Helpers\Digit as Helper;
  */
 class Digit extends Facade
 {
-    protected static function getFacadeAccessor()
+    protected static function getFacadeAccessor(): string
     {
         return Helper::class;
     }
