@@ -36,7 +36,7 @@ use SplFileInfo;
  */
 class Directory extends Facade
 {
-    protected static function getFacadeAccessor()
+    protected static function getFacadeAccessor(): string
     {
         return Helper::class;
     }
