@@ -35,7 +35,7 @@ use SplFileInfo;
  */
 class File extends Facade
 {
-    protected static function getFacadeAccessor()
+    protected static function getFacadeAccessor(): string
     {
         return Helper::class;
     }
