@@ -31,7 +31,7 @@ use DragonCode\Support\Helpers\OS as Helper;
  */
 class OS extends Facade
 {
-    protected static function getFacadeAccessor()
+    protected static function getFacadeAccessor(): string
     {
         return Helper::class;
     }
